requirements.txt
tensorflow>=2.11.0
opencv-python>=4.5.3
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.2
matplotlib>=3.4.2
pyswarm>=0.6
streamlit>=1.10.0
Pillow>=9.0.0